Update plugin with field provider support and more
This commit adds support for a field as a the image provider. It adds a config option for that.

The update also adds a small update for (globally for now) user defined title and caption elements. I find this useful as it allows me to set which of my fields I want to show.

Signed-off-by: Mike Wink <iam.mikewink@gmail.com>